+++ NEWS.md
@@ -1,6 +1,7 @@
 # spant 4.4.0
 * Fix for Philips svs spar/sdat voxel geometry.
 * Bug fixes for match_lineshape function.
+* Added zero_spec_threshold function.
 
 # spant 4.3.0
 * Bug fix for mr_data2bids function.